+++ san_francisco/src/main/java/org/eqasim/san_francisco/mode_choice/utilities/estimators/SanFranciscoWalkUtilityEstimator.java
@@ -1,0 +1,50 @@
+package org.eqasim.san_francisco.mode_choice.utilities.estimators;
+
+import java.util.List;
+
+import org.eqasim.core.simulation.mode_choice.utilities.estimators.WalkUtilityEstimator;
+import org.eqasim.core.simulation.mode_choice.utilities.predictors.PersonPredictor;
+import org.eqasim.core.simulation.mode_choice.utilities.predictors.WalkPredictor;
+import org.eqasim.san_francisco.mode_choice.parameters.SanFranciscoModeParameters;
+import org.eqasim.san_francisco.mode_choice.utilities.predictors.SanFranciscoPersonPredictor;
+import org.eqasim.san_francisco.mode_choice.utilities.variables.SanFranciscoPersonVariables;
+import org.matsim.api.core.v01.population.Person;
+import org.matsim.api.core.v01.population.PlanElement;
+import org.matsim.core.utils.geometry.CoordUtils;
+
+import com.google.inject.Inject;
+
+import ch.ethz.matsim.discrete_mode_choice.model.DiscreteModeChoiceTrip;
+
+public class SanFranciscoWalkUtilityEstimator extends WalkUtilityEstimator {
+	private final SanFranciscoModeParameters parameters;
+	private final SanFranciscoPersonPredictor predictor;
+
+	@Inject
+	public SanFranciscoWalkUtilityEstimator(SanFranciscoModeParameters parameters, PersonPredictor personPredictor,
+			WalkPredictor walkPredictor, SanFranciscoPersonPredictor predictor) {
+		super(parameters, walkPredictor);
+
+		this.parameters = parameters;
+		this.predictor = predictor;
+	}
+
+	protected double estimateRegionalUtility(SanFranciscoPersonVariables variables) {
+		return (variables.cityTrip) ? parameters.sfWalk.alpha_walk_city : 0.0;
+	}
+
+	@Override
+	public double estimateUtility(Person person, DiscreteModeChoiceTrip trip, List<? extends PlanElement> elements) {
+		SanFranciscoPersonVariables variables = predictor.predictVariables(person, trip, elements);
+
+		double utility = 0.0;
+		double distance = CoordUtils.calcEuclideanDistance(trip.getOriginActivity().getCoord(),
+				trip.getDestinationActivity().getCoord());
+		if (distance > 4 * 5280)
+			utility += -100;
+		utility += super.estimateUtility(person, trip, elements);
+		utility += estimateRegionalUtility(variables);
+
+		return utility;
+	}
+}