+++ san_francisco/src/main/java/org/eqasim/san_francisco/scenario/RunAdaptConfig.java
@@ -1,0 +1,34 @@
+package org.eqasim.san_francisco.scenario;
+
+import org.eqasim.core.components.config.ConfigAdapter;
+import org.eqasim.core.components.config.EqasimConfigGroup;
+import org.eqasim.core.simulation.EqasimConfigurator;
+import org.eqasim.san_francisco.mode_choice.SanFranciscoModeChoiceModule;
+import org.matsim.api.core.v01.TransportMode;
+import org.matsim.core.config.CommandLine.ConfigurationException;
+import org.matsim.core.config.Config;
+
+import ch.ethz.matsim.discrete_mode_choice.modules.config.DiscreteModeChoiceConfigGroup;
+
+public class RunAdaptConfig {
+	static public void main(String[] args) throws ConfigurationException {
+		ConfigAdapter.run(args, EqasimConfigurator.getConfigGroups(), RunAdaptConfig::adaptConfiguration);
+	}
+
+	static public void adaptConfiguration(Config config) {
+		// Ignore some input files
+		config.transit().setVehiclesFile(null);
+		config.households().setInputFile(null);
+
+		// Set up mode choice
+		EqasimConfigGroup eqasimConfig = EqasimConfigGroup.get(config);
+
+		eqasimConfig.setCostModel(TransportMode.car, SanFranciscoModeChoiceModule.CAR_COST_MODEL_NAME);
+		eqasimConfig.setCostModel(TransportMode.pt, SanFranciscoModeChoiceModule.PT_COST_MODEL_NAME);
+
+		DiscreteModeChoiceConfigGroup dmcConfig = (DiscreteModeChoiceConfigGroup) config.getModules()
+				.get(DiscreteModeChoiceConfigGroup.GROUP_NAME);
+
+		dmcConfig.setModeAvailability(SanFranciscoModeChoiceModule.MODE_AVAILABILITY_NAME);
+	}
+}