+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # CHANGELOG
-
+- Remove bike from Sao Paulo
+- Add Los Angeles and San Francisco
 - Add possibility to choose modes to RunPopulationRouting
 - Bugfix: EnrichedTransitRoutingModule wrote wrong link and facilities in "pt interaction" activities, which lead to problems when cutting the population
+- Bugfix: Added non-default estimators for car to SF and LA
+
 
 **1.0.5**